chore: Cell 클래스, 한 파일에 모아서 관리하도록 변경
