+++ src/main/kotlin/minesweeper/model/Cell.kt
@@ -23,3 +23,39 @@ abstract class Cell(
 
     abstract fun open(): Cell
 }
+
+class ClosedMine(
+    position: CellPosition,
+    override val isOpened: Boolean = false,
+    override val isMine: Boolean = true,
+) : Cell(position) {
+
+    override fun open(): Cell = OpenedMine(position)
+}
+
+class ClosedNonMine(
+    position: CellPosition,
+    override val isOpened: Boolean = false,
+    override val isMine: Boolean = false,
+) : Cell(position) {
+
+    override fun open(): Cell = OpenedNonMine(position)
+}
+
+class OpenedMine(
+    position: CellPosition,
+    override val isOpened: Boolean = true,
+    override val isMine: Boolean = true,
+) : Cell(position) {
+
+    override fun open(): Cell = this
+}
+
+class OpenedNonMine(
+    position: CellPosition,
+    override val isOpened: Boolean = true,
+    override val isMine: Boolean = false,
+) : Cell(position) {
+
+    override fun open(): Cell = this
+}
