--- src/main/kotlin/minesweeper/model/ClosedNonMine.kt
@@ -1,10 +1,0 @@
-package minesweeper.model
-
-class ClosedNonMine(
-    position: CellPosition,
-    override val isOpened: Boolean = false,
-    override val isMine: Boolean = false,
-) : Cell(position) {
-
-    override fun open(): Cell = OpenedNonMine(position)
-}
